land-title-diligence/backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Settings & validation
pydantic==2.9.0
pydantic-settings==2.5.2

# Supabase
supabase==2.7.4

# Anthropic / Claude
anthropic==0.34.2

# Embeddings
sentence-transformers==3.1.1
torch==2.4.1          # CPU build is fine; change to torch+cuda for GPU
transformers==4.44.2

# OCR — Tesseract
pytesseract==0.3.13
Pillow==10.4.0
PyMuPDF==1.24.10      # PDF → image (fitz)

# OCR — AWS Textract (optional)
boto3==1.35.20

# Utilities
python-jose[cryptography]==3.3.0
httpx==0.27.2
